Move fixtures into test folder
