+++ test/_setup.test.js
@@ -1,7 +1,7 @@
 import mocha from 'mocha';
 import sequelizeFixtures from 'sequelize-fixtures';
 import models from '../src/models';
-import seed, { initTokens } from '../fixtures/seed';
+import seed, { initTokens } from './fixtures/seed';
 
 describe('Test Setup', function() {
   it('should set up & seed database', function(done) {
@@ -9,7 +9,7 @@ describe('Test Setup', function() {
       models.sequelize
         .drop()
         .then(() => models.sequelize.sync())
-        .then(() => sequelizeFixtures.loadFile('fixtures/*.json', models))
+        .then(() => sequelizeFixtures.loadFile('test/fixtures/*.json', models))
         .then(() => initTokens())
         .then(() => done());
     } else {
