+++ test/fixtures/rooms.json
